Add justList property to only shows the data given
This property is useful when your suggestions are filtered on backend. What happen if someone has typo on Canada word? the default matching of the plugin won't show you the Canada country. That's the reason that you can filter your suggestion from backend and only use this component to show the options.